Move keyboard layout in its own class
