+++ gtk2_ardour/wscript
@@ -125,6 +125,7 @@ gtk2_ardour_sources = [
         'ghostregion.cc',
         'global_port_matrix.cc',
         'group_tabs.cc',
+        'keyboardlayout.cc',
         'pianokeyboard.cc',
         'gui_object.cc',
         'idleometer.cc',
